requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==7.0.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bot-studio==1.4.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver-binary==113.0.5672.63.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
fonttools==4.39.4
frozenlist==1.3.3
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
linkedin-api==2.0.1
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.5.1
plotly==5.15.0
pony==0.7.16
promise==2.3
pwinput==1.0.3
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-linkedin==4.1
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
Rx==1.6.3
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.1
shortuuid==1.0.11
singledispatch==3.7.0
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
